# Conflicts:
#	athena-gcs/pom.xml
#	athena-gcs/src/main/java/com/amazonaws/athena/connectors/gcs/GcsMetadataHandler.java
#	athena-gcs/src/main/java/com/amazonaws/athena/connectors/gcs/GcsRecordHandler.java
#	athena-gcs/src/main/java/com/amazonaws/athena/connectors/gcs/GcsSchemaUtils.java
#	athena-gcs/src/main/java/com/amazonaws/athena/connectors/gcs/GcsUtil.java
#	athena-gcs/src/main/java/com/amazonaws/athena/connectors/gcs/TypeUtils.java
#	athena-gcs/src/main/java/com/amazonaws/athena/connectors/gcs/storage/StorageMetadataImpl.java
#	athena-gcs/src/test/java/com/amazonaws/athena/connectors/gcs/GcsMetadataHandlerTest.java
#	athena-gcs/src/test/java/com/amazonaws/athena/connectors/gcs/GcsRecordHandlerTest.java
#	athena-gcs/src/test/java/com/amazonaws/athena/connectors/gcs/GcsSchemaUtilsTest.java
#	athena-gcs/src/test/java/com/amazonaws/athena/connectors/gcs/GcsUtilsTest.java
#	athena-gcs/src/test/java/com/amazonaws/athena/connectors/gcs/filter/FilterExpressionBuilderTest.java

+++ athena-gcs/src/main/java/com/amazonaws/athena/connectors/gcs/TypeUtils.java
@@ -61,15 +61,15 @@ public class TypeUtils
                 }
             case FLOAT8:
                 if (origVal instanceof Integer) {
-                    return ((Double) origVal).intValue();
+                    return (double) (int) origVal;
                 }
                 else if (origVal instanceof Float) {
-                    return ((Double) origVal).floatValue();
+                    return (double) (float) origVal;
                 }
                 return origVal;
             case FLOAT4:
                 if (origVal instanceof Integer) {
-                    return ((Float) origVal).intValue();
+                    return (float) (int) origVal;
                 }
                 else if (origVal instanceof Double) {
                     return ((Double) origVal).floatValue();
